Reorganize track-changes examples, update Move to support inline movements.
