+++ src/components/Section.ts
@@ -32,6 +32,8 @@ import { evaluateXPathToMap } from '../utilities/xquery.ts';
 import type { BookmarkRangeEnd } from './BookmarkRangeEnd.ts';
 import type { BookmarkRangeStart } from './BookmarkRangeStart.ts';
 import type { Move } from './Move.ts';
+import type { MoveRangeEnd } from './MoveRangeEnd.ts';
+import type { MoveRangeStart } from './MoveRangeStart.ts';
 import { Paragraph } from './Paragraph.ts';
 import type { Table } from './Table.ts';
 
@@ -43,13 +45,19 @@ export type SectionChild =
 	| Paragraph
 	| Table
 	| BookmarkRangeStart
-	| BookmarkRangeEnd;
+	| BookmarkRangeEnd
+	| MoveRangeStart
+	| MoveRangeEnd
+	| Move;
 
 export const sectionChildComponentNames = [
 	'Table',
 	'Paragraph',
 	'BookmarkRangeStart',
 	'BookmarkRangeEnd',
+	'MoveRangeStart',
+	'MoveRangeEnd',
+	'Move',
 ];
 
 /**
